+++ AngularClient/src/main/core/entities/element.spec.ts
@@ -1,6 +1,3 @@
-import { ElementFieldDataType } from "./element-field";
-import { TestHelpers } from "./test-helpers";
-
 describe("main/core/entities/element", () => {
 
 });