+++ AngularClient/src/main/core/entities/project.ts
@@ -1,9 +1,6 @@
-import { Subject } from "rxjs";
-
 import { Element } from "./element";
 import { EntityBase } from "./entity-base";
 import { User } from "./user";
-import { stripInvalidChars } from "../../shared/utils";
 
 export class Project extends EntityBase {